pyproject.toml
[tool.poetry]
name = "django-prac"
version = "0.1.0"
description = ""
authors = ["SaJH <wogur981208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-environ = "^0.11.2"
djangorestframework-camel-case = "^1.4.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.7.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.4.0"

factory-boy = "^3.3.0"
faker = "^19.6.1"

mypy = "^1.5.1"

types-pillow = "^10.1.0.2"
django-stubs = "^4.2.4"
djangorestframework-stubs = "^3.14.2"
drf-yasg-stubs = "^0.1.4"

flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^24.3.0"
pre-commit = "^3.4.0"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 140
target-version = ['py311']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "*/config/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
